修改数据库类型，price和mileage为float

+++ mysite/fetchcar/models.py
@@ -18,8 +18,8 @@ class Car(models.Model):
     city = models.CharField(max_length=30, null=True)
     car_type = models.ForeignKey(CarType, null=True)
     reg_time = models.DateField(null=True)
-    mileage = models.IntegerField(null=True)
-    price = models.IntegerField(null=True)
+    mileage = models.FloatField(null=True)
+    price = models.FloatField(null=True)
 
     def __unicode__(self):
         return "<{}\t{}\t{}>".format(self.brand, self.car_type, self.price)
